+++ src/main/java/ru/innopolis/stc9/controllers/EditUserController.java
@@ -2,7 +2,9 @@ package ru.innopolis.stc9.controllers;
 
 
 import org.apache.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import ru.innopolis.stc9.service.AdminService;
+import ru.innopolis.stc9.service.AdminServiceImpl;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,13 +13,14 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/views/edit-user")
+//@WebServlet("/views/edit-user")
 public class EditUserController extends HttpServlet{
     private final Logger logger = Logger.getLogger(GroupController.class);
     private static final String ENCODING = "UTF-8";
-    private final AdminService adminService = new AdminService();
+    //@Autowired
+    private final AdminService adminService = new AdminServiceImpl();
 
-    @Override
+    /*@Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         int id = Integer.parseInt(req.getParameter("user-id"));
         if (id>0) {
@@ -37,5 +40,5 @@ public class EditUserController extends HttpServlet{
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String msg = adminService.editUser(req);
         resp.sendRedirect(req.getContextPath() + msg);
-    }
+    }*/
 }
