+++ src/main/resources/log4j.properties
@@ -4,3 +4,5 @@ log4j.appender.defaultLog=org.apache.log4j.ConsoleAppender
 log4j.appender.defaultLog.Target=System.out
 log4j.appender.defaultLog.layout=org.apache.log4j.PatternLayout
 log4j.appender.defaultLog.layout.conversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
+
+logging.level.org.springframework.security=DEBUG
