+++ src/main/webapp/WEB-INF/web.xml
@@ -59,32 +59,4 @@
         <url-pattern>/*</url-pattern>
     </filter-mapping>
 
-<!--next need to delete***********************************************************************-->
-    <!-- <servlet>
-         <servlet-name>UsersListController</servlet-name>
-         <servlet-class>ru.innopolis.stc9.controllers.UsersListController</servlet-class>
-     </servlet>
-     <servlet-mapping>
-         <servlet-name>UsersListController</servlet-name>
-         <url-pattern>/users-list/</url-pattern>
-     </servlet-mapping>
-
-     <servlet>
-         <servlet-name>EditUserController</servlet-name>
-         <servlet-class>ru.innopolis.stc9.controllers.EditUserController</servlet-class>
-     </servlet>
-     <servlet-mapping>
-         <servlet-name>EditUserController</servlet-name>
-         <url-pattern>/edit-user</url-pattern>
-     </servlet-mapping>
-
-     <filter>
-         <filter-name>auth-filter</filter-name>
-         <filter-class>ru.innopolis.stc9.controllers.filter.UserFilter</filter-class>
-     </filter>
-     <filter-mapping>
-         <filter-name>auth-filter</filter-name>
-         <url-pattern>/views/*</url-pattern>
-     </filter-mapping>-->
-
 </web-app>
