+++ src/main/webapp/views/editUser.jsp
@@ -8,44 +8,44 @@
                 <div class="panel-title ">Редактирование пользователя</div>
             </div>
             <div class="content-box-large box-with-header">
-                <form class="form-horizontal" action="${pageContext.request.contextPath}/views/edit-user" method="post">
+                <form class="form-horizontal" action="${pageContext.request.contextPath}/admin/edit-user" method="post">
                     <fieldset>
                         <legend>Основные данные</legend>
                         <div class="form-group">
                             <label for="userId" class="col-sm-2 control-label">Id</label>
                             <div class="col-sm-10"><input type="text" class="form-control" id="userId"
                                                           placeholder="User ID"
-                                                          name="userId"
+                                                          name="id"
                                                           value="${user.id}" readonly></div>
                         </div>
                         <div class="form-group">
                             <label for="editLogin" class="col-sm-2 control-label">Логин</label>
                             <div class="col-sm-10"><input type="text" class="form-control" id="editLogin"
-                                                          placeholder="Login" name="editLogin"
+                                                          placeholder="Login" name="login"
                                                           value="${user.login}"></div>
                         </div>
                         <div class="form-group">
                             <label for="editFirstName" class="col-sm-2 control-label">Имя</label>
                             <div class="col-sm-10"><input type="text" class="form-control" id="editFirstName"
                                                           placeholder="First name"
-                                                          name="editFirstName" value="${user.firstName}"></div>
+                                                          name="first_name" value="${user.firstName}"></div>
                         </div>
                         <div class="form-group">
                             <label for="editLastName" class="col-sm-2 control-label">Фамилия</label>
                             <div class="col-sm-10"><input type="text" class="form-control" id="editLastName"
                                                           placeholder="Last name"
-                                                          name="editLastName" value="${user.secondName}"></div>
+                                                          name="second_name" value="${user.secondName}"></div>
                         </div>
                         <div class="form-group">
                             <label for="editMiddleName" class="col-sm-2 control-label">Отчество</label>
                             <div class="col-sm-10"><input type="text" class="form-control" id="editMiddleName"
                                                           placeholder="Middle name"
-                                                          name="editMiddleName" value="${user.middleName}"></div>
+                                                          name="middle_name" value="${user.middleName}"></div>
                         </div>
                         <div class="form-group">
                             <label class="col-sm-2 control-label">Права доступа</label>
                             <div class="col-sm-10">
-                                <select class="form-control" name="editRole">
+                                <select class="form-control" name="permission_group">
                                     <option disabled>Необходимо выбрать роль пользователя</option>
                                     <option ${user.permissionGroup == 'ROLE_ADMIN' ? "selected" : ""} value="ROLE_ADMIN">Admin
                                     </option>
@@ -59,7 +59,7 @@
                         <div class="form-group">
                             <label class="col-sm-2 control-label">Включение/Отключение аккаунта</label>
                             <div class="col-sm-10">
-                                <select class="form-control" name="editEnabled">
+                                <select class="form-control" name="enabled">
                                     <option disabled>Выбирите бедт ли активен аккаунт</option>
                                     <option ${user.enabled == 1 ? "selected" : ""} value="1">Активен
                                     </option>
@@ -75,7 +75,7 @@
                             <label for="oldPassword" class="col-sm-2 control-label">Старый пароль</label>
                             <div class="col-sm-10"><input class="form-control" type="password" id="oldPassword"
                                                           plaseholder="Old password"
-                                                          name="oldPassword" value=""></div>
+                                                          name="hash_password" value=""></div>
                         </div>
                         <div class="form-group">
                             <label for="newPassword" class="col-sm-2 control-label">Новый пароль</label>
@@ -96,9 +96,9 @@
                         </div>
                     </div>
 
-                    <div>${userUpdateMessage}</div>
-                    <div>${loginUpdateMessage}</div>
-                    <c:if test="${passwordUpdateMessage != null}"><div class='alert alert-danger'>${passwordUpdateMessage}</div></c:if>
+                    <c:if test="${errors != null}">
+                        <c:forEach items="${errors}" var="error"><div class="alert alert-success">${error}</div></c:forEach>
+                    </c:if>
 
                 </form>
                 <br/><br/>
